+++ src/dgisim/character/character.py
@@ -61,6 +61,7 @@ __all__ = [
     "MaguuKenki",
     "Mona",
     "Nahida",
+    "Nilou",
     "Ningguang",
     "Noelle",
     "Qiqi",
@@ -3173,6 +3174,90 @@ class Nahida(Character):
                 status=stt.ShrineOfMayaStatus,
             ))
         return tuple(effects)
+
+    @classmethod
+    def from_default(cls, id: int = -1) -> Self:
+        return cls(
+            id=id,
+            alive=True,
+            hp=10,
+            max_hp=10,
+            energy=0,
+            max_energy=2,
+            hiddens=stts.Statuses(()),
+            statuses=stts.Statuses(()),
+            elemental_aura=ElementalAura.from_default(),
+        )
+
+
+class Nilou(Character):
+    _ELEMENT = Element.HYDRO
+    _WEAPON_TYPE = WeaponType.SWORD
+    _TALENT_STATUS = stt.TheStarrySkiesTheirFlowersRainStatus
+    _FACTIONS = frozenset((Faction.SUMERU,))
+
+    _SKILL1_COST = AbstractDice({
+        Element.HYDRO: 1,
+        Element.ANY: 2,
+    })
+    _SKILL2_COST = AbstractDice({
+        Element.HYDRO: 3,
+    })
+    _ELEMENTAL_BURST_COST = AbstractDice({
+        Element.HYDRO: 3,
+    })
+
+    @override
+    def _skill1(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        return normal_attack_template(
+            game_state=game_state,
+            source=source,
+            element=Element.PHYSICAL,
+            damage=2,
+        )
+
+    @override
+    def _skill2(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        effects: list[eft.Effect] = [
+            eft.ReferredDamageEffect(
+                source=source,
+                target=DynamicCharacterTarget.OPPO_ACTIVE,
+                element=Element.HYDRO,
+                damage=3,
+                damage_type=DamageType(elemental_skill=True),
+            ),
+        ]
+        hydro_dendro_char_only = all([
+            char.ELEMENT in (Element.HYDRO, Element.DENDRO)
+            for char in game_state.get_player(source.pid).characters
+        ])
+        if hydro_dendro_char_only:
+            effects.append(eft.AddCombatStatusEffect(
+                target_pid=source.pid,
+                status=stt.GoldenChalicesBountyStatus,
+            ))
+        return tuple(effects)
+
+    @override
+    def _elemental_burst(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        return (
+            eft.EnergyDrainEffect(
+                target=source,
+                amount=self.max_energy,
+            ),
+            eft.ReferredDamageEffect(
+                source=source,
+                target=DynamicCharacterTarget.OPPO_ACTIVE,
+                element=Element.HYDRO,
+                damage=2,
+                damage_type=DamageType(elemental_burst=True),
+            ),
+            eft.RelativeAddCharacterStatusEffect(
+                source_pid=source.pid,
+                target=DynamicCharacterTarget.OPPO_ACTIVE,
+                status=stt.LingeringAeonStatus,
+            ),
+        )
 
     @classmethod
     def from_default(cls, id: int = -1) -> Self: